pyproject.toml
[project]
name = "youtube-chatbot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "chromadb>=1.4.1",
    "flask>=3.1.2",
    "flask-cors>=6.0.2",
    "huggingface-hub>=0.36.0",
    "langchain>=1.2.6",
    "langchain-community>=0.4.1",
    "langchain-core>=1.2.7",
    "langchain-huggingface>=1.2.0",
    "langchain-text-splitters>=1.1.0",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
    "sentence-transformers>=5.2.0",
    "youtube-transcript-api>=1.2.3",
]
